Start of big refactor
This is the first step of a big refactor + functionality leap.
The step is split out in order to ensure git properly keeps track of the
change history.
