+++ src/main/java/com/insightfullogic/honest_profiler/ports/javafx/JavaFXApplication.java
@@ -27,14 +27,14 @@ import org.picocontainer.PicoBuilder;
 import com.insightfullogic.honest_profiler.core.Monitor;
 import com.insightfullogic.honest_profiler.core.filters.ProfileFilter;
 import com.insightfullogic.honest_profiler.ports.LoggerInjector;
+import com.insightfullogic.honest_profiler.ports.javafx.controller.FlameGraphViewModel;
+import com.insightfullogic.honest_profiler.ports.javafx.controller.FlatViewModel;
+import com.insightfullogic.honest_profiler.ports.javafx.controller.ProfileViewModel;
+import com.insightfullogic.honest_profiler.ports.javafx.controller.TraceCountViewModel;
+import com.insightfullogic.honest_profiler.ports.javafx.controller.TreeTableViewModel;
 import com.insightfullogic.honest_profiler.ports.javafx.landing.LandingViewModel;
 import com.insightfullogic.honest_profiler.ports.javafx.profile.CachingProfileListener;
-import com.insightfullogic.honest_profiler.ports.javafx.profile.FlameGraphViewModel;
-import com.insightfullogic.honest_profiler.ports.javafx.profile.FlatViewModel;
 import com.insightfullogic.honest_profiler.ports.javafx.profile.ProfileListenerProvider;
-import com.insightfullogic.honest_profiler.ports.javafx.profile.ProfileViewModel;
-import com.insightfullogic.honest_profiler.ports.javafx.profile.TraceCountViewModel;
-import com.insightfullogic.honest_profiler.ports.javafx.profile.TreeTableViewModel;
 import com.insightfullogic.honest_profiler.ports.sources.LocalMachineSource;
 import com.insightfullogic.honest_profiler.ports.web.store.FileLogRepository;
