+++ src/main/resources/com/insightfullogic/honest_profiler/ports/javafx/FlameGraphView.fxml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <?import com.insightfullogic.honest_profiler.ports.javafx.flame_graph.FlameGraphCanvas?>
-<FlameGraphCanvas fx:controller="com.insightfullogic.honest_profiler.ports.javafx.profile.FlameGraphViewModel"
+<FlameGraphCanvas fx:controller="com.insightfullogic.honest_profiler.ports.javafx.controller.FlameGraphViewModel"
                   fx:id="flameView"
                   xmlns="http://javafx.com/javafx/8"
                   xmlns:fx="http://javafx.com/fxml/1">
