+++ src/main/resources/com/insightfullogic/honest_profiler/ports/javafx/FlatProfileView.fxml
@@ -2,7 +2,7 @@
 
 <?import javafx.scene.control.TableColumn?>
 <?import javafx.scene.control.TableView?>
-<TableView fx:controller="com.insightfullogic.honest_profiler.ports.javafx.profile.FlatViewModel"
+<TableView fx:controller="com.insightfullogic.honest_profiler.ports.javafx.controller.FlatViewModel"
            fx:id="flatProfileView" prefHeight="9999" prefWidth="640.0" xmlns="http://javafx.com/javafx/8"
            xmlns:fx="http://javafx.com/fxml/1">
     <columns>
