+++ src/main/resources/com/insightfullogic/honest_profiler/ports/javafx/ProfileView.fxml
@@ -4,7 +4,7 @@
 <?import javafx.scene.layout.StackPane?>
 <?import javafx.scene.layout.VBox?>
 <VBox prefHeight="400.0" prefWidth="640.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1"
-      fx:controller="com.insightfullogic.honest_profiler.ports.javafx.profile.ProfileViewModel">
+      fx:controller="com.insightfullogic.honest_profiler.ports.javafx.controller.ProfileViewModel">
     <children>
         <ToolBar prefWidth="200.0">
             <items>
