+++ src/main/resources/com/insightfullogic/honest_profiler/ports/javafx/TraceCountView.fxml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <?import javafx.scene.control.Label?>
-<Label fx:controller="com.insightfullogic.honest_profiler.ports.javafx.profile.TraceCountViewModel"
+<Label fx:controller="com.insightfullogic.honest_profiler.ports.javafx.controller.TraceCountViewModel"
        fx:id="traceCount"
        prefHeight="15.0"
        prefWidth="104.0"
